Moving SIMPLRestClient out of SIMPLib into its own library.
