+++ Source/RESTClient/RESTClientConfiguration.h.in
@@ -1,0 +1,40 @@
+/* This file is auto generated. DO NOT EDIT THIS FILE. Edit the original template file and then rerun CMake
+* again to regenerate this file.
+*/
+
+#ifndef _RESTClient_H_
+#define _RESTClient_H_
+
+
+/* Define to 1 if RESTClient was built as a dynamic library */
+#cmakedefine RESTClient_BUILT_AS_DYNAMIC_LIB @RESTClient_BUILT_AS_DYNAMIC_LIB@
+
+/* Include the Overall Configuration header file */
+#include "@PROJECT_NAME@/@CMP_CONFIGURATION_FILE_NAME@"
+
+/* Include the Typedefs Header */
+#include "@PROJECT_NAME@/@CMP_TYPES_FILE_NAME@"
+
+/* Include the DLL export preprocessor defines */
+#include "@PROJECT_NAME@/@PROJECT_NAME@DLLExport.h"
+
+/* We are going to disable some warnings just to get a grip on all the warnings that get produced
+* during an MSVC build */
+#cmakedefine DREAM3D_DISABLE_MSVC_WARNINGS
+
+#define RESTClient_USE_MUTEX
+
+#if (_MSC_VER >= 1)
+
+  #pragma warning (error: 4715) /* Not all control points return a value */
+
+  #ifdef DREAM3D_DISABLE_MSVC_WARNINGS
+    #pragma warning(disable:4244)
+    #pragma warning(disable:4267)
+    #pragma warning(disable:4305)
+  #endif
+
+#endif
+
+
+#endif
