+++ Source/RESTClient/SIMPLClientRequest.cpp
