+++ Source/RESTClient/mrestrequestqueue.h
@@ -23,37 +23,13 @@ SOFTWARE.
 
 #pragma once
 
-#include <QByteArray>
-#include <QHash>
-#include <QMetaType>
+#include "mrestrequestptr.h"
+#include <QQueue>
 
-#define CONFIG_VALUE(name, type) mValues.insert(#name, ValuePtr(type, static_cast<void*>(&name)));
+#include "RESTClient/RESTClient.h"
 
-class Config
+class RESTClient_EXPORT MRestRequestQueue : public QQueue<MRestRequestPtr>
 {
 public:
-  Config(const QByteArray& groupName);
-  void load();
-  void save();
-
-protected:
-  class ValuePtr
-  {
-  public:
-    ValuePtr()
-    {
-    }
-    ValuePtr(int t, void* v)
-    : type(t)
-    , ptr(v)
-    {
-    }
-    int type = QMetaType::UnknownType;
-    void* ptr = nullptr;
-  };
-  QHash<QByteArray, ValuePtr> mValues;
-
-private:
-  const QByteArray mGroupName;
-  static void copyValue(void* dst, int type, const QVariant& value);
+  void enqueue(MRestRequestPtr request);
 };
